#1472 - Solved user rights issues by passing a list of case ids to the service method and using that list to restrict the resulting samples

+++ sormas-backend/src/main/java/de/symeda/sormas/backend/common/AbstractAdoService.java
@@ -175,10 +175,10 @@ public abstract class AbstractAdoService<ADO extends AbstractDomainObject> imple
 		return em.createQuery(cq).getResultList();
 	}
 	
-	public List<Integer> getAllIds(User user) {
+	public List<Long> getAllIds(User user) {
 		
 		CriteriaBuilder cb = em.getCriteriaBuilder();
-		CriteriaQuery<Integer> cq = cb.createQuery(Integer.class);
+		CriteriaQuery<Long> cq = cb.createQuery(Long.class);
 		Root<ADO> from = cq.from(getElementClass());
 
 		if (user != null) {
